minimal_requirements.txt
attrs==22.1.0
cudatoolkit==11.7 # conda install cudatoolkit=11.7 -c pytorch -c nvidia
cudnn==8.9.2.26
numpy==1.23.4
Pillow==9.3.0
python==3.9.15 # Do this first
PyYAML==6.0
scikit_learn==1.1.3
scipy==1.10.0
torch==2.0.1 # conda install pytorch torchvision torchaudio pytorch-cuda=11.7 -c pytorch -c nvidia
torchmetrics==1.0.3
torchvision==0.15.2
tqdm==4.65.0
wandb==0.13.5